Fixed adding goods to the warehouse

+++ KS.ViewModels/Stock/StockIndexViewModel.cs
@@ -3,11 +3,11 @@
     public class StockIndexViewModel
     {
         public int Id { get; set; }
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
 
         public Entities.Product Product { get; set; }
 
-        public long WarehouseId { get; set; }
+        public int WarehouseId { get; set; }
 
         public Entities.Warehouse Warehouse { get; set; }
